Create base for profile view and update: request + response file, connecting the screens in nav_graph.xml, service
