+++ app/src/main/java/com/example/e_commerce/network/model/response/CurrentUserResponse.java
@@ -1,0 +1,5 @@
+package com.example.e_commerce.network.model.response;
+
+public class CurrentUserResponse {
+    // add info here later
+}
