+++ app/src/main/java/com/example/e_commerce/network/service/ProfileService.java
@@ -1,0 +1,20 @@
+package com.example.e_commerce.network.service;
+// create profile request + response and add them here later
+
+// import com.example.e_commerce.network.model.request.SearchProductRequest;
+import com.example.e_commerce.network.model.response.ResponseAPI;
+// import com.example.e_commerce.network.model.response.product.GetProductResponse;
+// import com.example.e_commerce.network.model.response.product.ProductResponse;
+// import com.example.e_commerce.network.model.response.product.TypeProductResponse;
+
+import java.util.List;
+
+import retrofit2.Call;
+import retrofit2.http.Body;
+import retrofit2.http.GET;
+import retrofit2.http.POST;
+import retrofit2.http.Query;
+
+public interface ProfileService {
+    // add API info later, with "user/info" (GET) and "user/update" (POST)
+}
