+++ app/src/main/java/com/example/e_commerce/screens/profile/ProfileEditFragment.java
@@ -1,4 +1,4 @@
-package com.example.e_commerce.screens;
+package com.example.e_commerce.screens.profile;
 
 import static androidx.navigation.Navigation.findNavController;
 
@@ -11,25 +11,23 @@ import android.view.View;
 import android.view.ViewGroup;
 
 import com.example.e_commerce.R;
-import com.example.e_commerce.databinding.FragmentProfileBinding;
-import com.example.e_commerce.network.model.request.CurrrentUserRequest;
+import com.example.e_commerce.databinding.FragmentProfileEditBinding;
 
-import javax.inject.Inject;
+public class ProfileEditFragment extends Fragment {
 
-public class ProfileFragment extends Fragment {
-
-    private FragmentProfileBinding binding;
-
+   private FragmentProfileEditBinding binding;
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        binding = FragmentProfileBinding.inflate(inflater, container, false);
-        binding.editUserInfo.setOnClickListener(new View.OnClickListener() {
+        binding = FragmentProfileEditBinding.inflate(inflater, container, false);
+        binding.updateUserInfo.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                findNavController(getView()).navigate(R.id.action_profileFragment_to_profileEditFragment);
+                Bundle bundle = new Bundle();
+                // needs to add User details here
+                findNavController(getView()).navigate(R.id.action_profileEditFragment_to_profileFragment, bundle);
             }
         });
         return binding.getRoot();
