+++ app/src/main/java/com/example/e_commerce/screens/profile/ProfileFragment.java
@@ -1,0 +1,37 @@
+package com.example.e_commerce.screens.profile;
+
+import static androidx.navigation.Navigation.findNavController;
+
+import android.os.Bundle;
+
+import androidx.fragment.app.Fragment;
+
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import com.example.e_commerce.R;
+import com.example.e_commerce.databinding.FragmentProfileBinding;
+import com.example.e_commerce.network.model.request.CurrrentUserRequest;
+
+import javax.inject.Inject;
+
+public class ProfileFragment extends Fragment {
+
+    private FragmentProfileBinding binding;
+
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        // Inflate the layout for this fragment
+        binding = FragmentProfileBinding.inflate(inflater, container, false);
+        binding.editUserInfo.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                findNavController(getView()).navigate(R.id.action_profileFragment_to_profileEditFragment);
+            }
+        });
+        return binding.getRoot();
+    }
+}
