+++ app/src/main/res/layout/fragment_profile.xml
@@ -5,8 +5,9 @@
     android:id="@+id/constraint"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".screens.ProfileFragment">
+    tools:context=".screens.profile.ProfileFragment">
 
+<!--All info is hard-coded, will come back later-->
     <!-- Profile image-->
     <ImageView
         android:id="@+id/imageView"
