+++ app/src/main/res/layout/fragment_profile_edit.xml
@@ -5,7 +5,7 @@
     android:id="@+id/constraint"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".screens.ProfileFragment">
+    tools:context=".screens.profile.ProfileFragment">
 
     <!-- Profile image-->
     <ImageView
@@ -122,6 +122,7 @@
         app:layout_constraintTop_toBottomOf="@id/address"/>
 
     <Button
+        android:id="@+id/update_user_info"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:text="Cập nhật"
