+++ app/src/main/res/navigation/nav_graph.xml
@@ -35,7 +35,7 @@
         tools:layout="@layout/fragment_chat" />
     <fragment
         android:id="@+id/profileFragment"
-        android:name="com.example.e_commerce.screens.ProfileFragment"
+        android:name="com.example.e_commerce.screens.profile.ProfileFragment"
         android:label="fragment_profile"
         tools:layout="@layout/fragment_profile" >
         <action
@@ -49,7 +49,11 @@
         android:label="ProductDetailFragment" />
     <fragment
         android:id="@+id/profileEditFragment"
-        android:name="com.example.e_commerce.ProfileEditFragment"
+        android:name="com.example.e_commerce.screens.profile.ProfileEditFragment"
         android:label="ProfileEditFragment"
-        tools:layout="@layout/fragment_profile_edit"/>
+        tools:layout="@layout/fragment_profile_edit">
+        <action
+            android:id="@+id/action_profileEditFragment_to_profileFragment"
+            app:destination="@id/profileFragment" />
+    </fragment>
 </navigation>
